+++ test/state.test.mjs
@@ -1,10 +1,9 @@
 import { expect } from 'chai';
-import { Mind, State, Belief, Subject, Archetype, Traittype, save_mind, load } from '../public/worker/cosmos.mjs';
+import { Mind, State, Belief, Subject, Archetype, Traittype, save_mind, load , logos } from '../public/worker/cosmos.mjs';
 import * as Cosmos from '../public/worker/cosmos.mjs';
 import * as DB from '../public/worker/db.mjs';
 import { createMindWithBeliefs, setupMinimalArchetypes, setupStandardArchetypes, get_first_belief_by_label } from './helpers.mjs';
 
-const logos = () => DB.get_logos_mind();
 
 describe('State', () => {
   beforeEach(() => {