+++ test/trait_operations.test.mjs
@@ -1,10 +1,9 @@
 import { expect } from 'chai'
-import { Mind, State, Belief, Archetype, Traittype } from '../public/worker/cosmos.mjs'
+import { Mind, State, Belief, Archetype, Traittype , logos } from '../public/worker/cosmos.mjs'
 import * as DB from '../public/worker/db.mjs'
 import {log} from '../public/lib/debug.mjs'
 import { stdTypes } from './helpers.mjs'
 
-const logos = () => DB.get_logos_mind();
 
 describe('Trait Operations Pattern', () => {
   beforeEach(() => {